Added polygons property to Feature
